Refactor get_gold into screen_helpers
